work/final/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
dotenv==0.9.9
Flask==3.1.0
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pycparser==2.22
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
WTForms==3.2.1
